+++ backend/src/main/java/com/debede/backend/repository/Carta_TCGRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Carta_TCGRepository extends JpaRepository<Carta_TCG, Long> {
+public interface Carta_TCGRepository extends JpaRepository<Carta_TCG, Integer> {
 
 }
